requirements.txt
django
django-htmx
django-simple-history
django-filter
django-simple-history